fix(quotation): change quantity from decimal to integer and fix update validation
- Change quantity field type from decimal to integer in model, migration, and casts
- Add QuotationUpdateRequest with proper validation rules for updates
- Fix date format validation to accept both d/m/Y and Y-m-d formats
- Make exchange_rate optional for normal quotations and required only for via type
- Include revision ID in form data and JavaScript initialization
- Add comprehensive test coverage for quotation updates
- Fix package-lock.json peer dependency issues

+++ beayar-erp/database/migrations/2026_01_25_085005_create_quotation_products_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string('unit')->nullable();
             $table->string('delivery_time')->nullable();
             $table->string('requision_no')->nullable();
-            $table->decimal('quantity', 10, 2);
+            $table->integer('quantity');
             $table->decimal('unit_price', 15, 2);
             $table->decimal('foreign_currency_buying', 15, 2)->nullable();
             $table->decimal('bdt_buying', 15, 2)->nullable();
